services/data-tunnel/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
sqlalchemy>=1.4
psycopg[binary]==3.2.3
pgvector==0.3.4
boto3==1.35.14
opensearch-py==2.6.0
celery==5.4.0
redis==5.0.8
nats-py==2.8.0
langdetect==1.0.9
dateparser==1.2.0
python-magic==0.4.27
pytesseract==0.3.10
Pillow==10.4.0
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
great-expectations==1.4.0
python-multipart>=0.0.9
unstructured>=0.14.6
pdfplumber>=0.11.0
pypdf>=4.2.0
python-docx>=1.1.2
openpyxl>=3.1.5
python-pptx>=0.6.23
beautifulsoup4>=4.12.2
lxml>=5.1.0
ocrmypdf>=16.3.0
